Updated search criteria
While working on a way to produce better fuzzy results, where a listing may refer to something for a range of products, I decided to reread the challenge rules for more clarity.  I then noticed this line:

A single price listing may match at most one product.

For most products, this isn't an issue.  The listing will either contain both the manufacturer and the model, or it won't.  For a listing like this:

{"title":"Duragadget Carry Case and Belt Clip for Fujifilm Finepix Series","manufacturer":"DURAGADGET","currency":"CAD","price":"22.49"}

This is where things get interesting.  This listing may match any product that contains the manufacturer and the family if family was part of the search.  Since this listing would potentially match 2 or more products, then according to the rules this listing could only be used once.  And since we're going for precision over quantity, I'd prefer to not match this type of record against anything.  Otherwise, if the search were to include ju