Refactoring indexing, adding filter words
While investigating the data further, I noticed that some 3rd party listings would mention the model and family, but not the manufacturer.  Since family is not part of the listings file, I had to parse and index the products file separately.  This now means that as long as we find a listing that contains either the manufacturer and model or the family and model, it will count as a hit.  

In the process of adding the new index, I also realized that some common words were getting indexed, creating vast amounts of useless data.  I added common word filtering to all fields, in order to mitigate the issue.

All of this resulted in increased accuracy.  Manufacturers are now cross referenced with titles and family, and common words are now filtered which reduces memory usage and some processing time.  Unfortunately processing times have tripled, mostly due to indexing.  Once everything is indexed, the actual search and results file creation only takes around 0.7 second